Created a custom header and removed unecessary imports
