+++ app/(protected)/_layout.tsx
@@ -1,76 +1,19 @@
-import { Tabs } from 'expo-router';
-//import { SafeAreaProvider } from "react-native-safe-area-context";
-import React, { MouseEventHandler } from 'react';
-//import * as eva from "@eva-design/eva";
-//import { ApplicationProvider } from "@ui-kitten/components";
-import AntDesign from '@expo/vector-icons/AntDesign';
-import Ionicons from '@expo/vector-icons/Ionicons';
-import { StyleSheet, View } from 'react-native';
-import { handleSignOut } from '../../supabase/supabase-signout';
-
-import {
-  TopNavigationAction,
-  Icon,
-  IconElement,
-  IconProps,
-  OverflowMenu,
-  MenuItem,
-  TopNavigation,
-  Layout,
-} from '@ui-kitten/components';
-import { SafeAreaView } from 'react-native-safe-area-context';
+import { Tabs } from "expo-router";
+import React, { MouseEventHandler } from "react";
+import AntDesign from "@expo/vector-icons/AntDesign";
+import Ionicons from "@expo/vector-icons/Ionicons";
+import { StyleSheet, View } from "react-native";
+import CustomHeader from "@/components/CustomHeader";
 
 export default function RootLayout() {
-  const [menuVisible, setMenuVisible] = React.useState(false);
-  const [disableSignOutButton, setDisableSignOutButton] = React.useState(false);
-  const toggleMenu = (): void => {
-    setMenuVisible(!menuVisible);
-  };
-
-  async function signOut() {
-    setDisableSignOutButton(true);
-    await handleSignOut();
-    setDisableSignOutButton(false);
-  }
-  const MenuIcon = (props: IconProps): IconElement => (
-    <Icon {...props} name="more-vertical" />
-  );
-
-  const LogOutIcon = (props: IconProps): IconElement => (
-    <Icon {...props} name="log-out" />
-  );
-
-  const renderMenuAction = (): React.ReactElement => (
-    <TopNavigationAction icon={MenuIcon} onPress={toggleMenu} />
-  );
-  const renderOverflowMenuAction = (): React.ReactElement => (
-    <OverflowMenu
-      anchor={renderMenuAction}
-      visible={menuVisible}
-      onBackdropPress={toggleMenu}
-      placement="bottom end"
-    >
-      <MenuItem accessoryLeft={LogOutIcon} title="Logout" onPress={signOut} />
-    </OverflowMenu>
-  );
   return (
-    // <ApplicationProvider {...eva} theme={eva.light}>
-    //<SafeAreaProvider>
-
     <React.Fragment>
-      <View>
-        <TopNavigation
-          accessoryRight={renderOverflowMenuAction}
-          style={styles.container}
-        />
-      </View>
-      {/* <View> */}
-      <Tabs>
+      <Tabs screenOptions={{ header: () => <CustomHeader /> }}>
         <Tabs.Screen
           name="index"
           options={{
-            title: 'Home',
-            tabBarLabel: 'Home',
+            title: "Home",
+            tabBarLabel: "Home",
             tabBarIcon: ({ color, size }) => (
               <AntDesign name="home" size={24} color="black" />
             ),
@@ -80,8 +23,8 @@ export default function RootLayout() {
         <Tabs.Screen
           name="dashboard"
           options={{
-            title: 'Dashboard',
-            tabBarLabel: 'Dashboard',
+            title: "Dashboard",
+            tabBarLabel: "Dashboard",
             tabBarIcon: ({ color, size }) => (
               <AntDesign name="dashboard" size={24} color="black" />
             ),
@@ -91,8 +34,8 @@ export default function RootLayout() {
         <Tabs.Screen
           name="library"
           options={{
-            title: 'Library',
-            tabBarLabel: 'Library',
+            title: "Library",
+            tabBarLabel: "Library",
             tabBarIcon: ({ color, size }) => (
               <Ionicons name="library-outline" size={24} color="black" />
             ),
@@ -102,28 +45,14 @@ export default function RootLayout() {
         <Tabs.Screen
           name="new-session"
           options={{
-            title: 'New Session',
-            tabBarLabel: 'New Session',
+            title: "New Session",
+            tabBarLabel: "New Session",
             tabBarIcon: ({ color, size }) => (
               <Ionicons name="create-outline" size={24} color="black" />
             ),
           }}
         />
       </Tabs>
-      {/* </View> */}
     </React.Fragment>
-
-    // </SafeAreaProvider>
-    //</ApplicationProvider>
   );
 }
-
-const styles = StyleSheet.create({
-  container: {
-    backgroundColor: '#FF0000',
-    marginTop: '10%',
-    marginBottom: '0%',
-    top: 5,
-    position: 'relative',
-  },
-});
